fix: syscall type

+++ eth-riscv-runtime/src/block.rs
@@ -9,7 +9,7 @@ pub fn timestamp() -> U256 {
     let third: u64;
     let fourth: u64;
     unsafe {
-        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u32::from(Syscall::Timestamp));
+        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u8::from(Syscall::Timestamp));
     }
     U256::from_limbs([first, second, third, fourth])
 }
@@ -21,7 +21,7 @@ pub fn base_fee() -> U256 {
     let third: u64;
     let fourth: u64;
     unsafe {
-        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u32::from(Syscall::BaseFee));
+        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u8::from(Syscall::BaseFee));
     }
     U256::from_limbs([first, second, third, fourth])
 }
@@ -30,7 +30,7 @@ pub fn base_fee() -> U256 {
 pub fn chain_id() -> u64 {
     let id: u64;
     unsafe {
-        asm!("ecall", lateout("a0") id, in("t0") u32::from(Syscall::ChainId));
+        asm!("ecall", lateout("a0") id, in("t0") u8::from(Syscall::ChainId));
     }
     id
 }
@@ -42,7 +42,7 @@ pub fn gas_limit() -> U256 {
     let third: u64;
     let fourth: u64;
     unsafe {
-        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u32::from(Syscall::GasLimit));
+        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u8::from(Syscall::GasLimit));
     }
     U256::from_limbs([first, second, third, fourth])
 }
@@ -54,7 +54,7 @@ pub fn number() -> U256 {
     let third: u64;
     let fourth: u64;
     unsafe {
-        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u32::from(Syscall::Number));
+        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, lateout("a3") fourth, in("t0") u8::from(Syscall::Number));
     }
     U256::from_limbs([first, second, third, fourth])
 }
